Correct typos and errors on S_n.
Define support of a permutation.
Try to get consistency when denoting a permutation. Maybe we should stick with Greek letters - I was used to use them.
Conjugation classes for S_4: I found only 3 2x2-cycles, which is consistent with the fact that |S_4| = 24 (and not 27).
Some corrections with permutations in proofs.